requirements.txt
# Core dependencies
prefect
tweepy
openai>=1.0.0
PyGithub>=2.1.1
python-dotenv # From experimental
crewai>=0.1,<0.2  # For CrewAI adapter

# Langchain related (from experimental)
langchain-core 
langchain-community
langchain-openai

# ChromaDB and Embeddings
chromadb>=1.0.0
sentence-transformers>=2.2.2

# Testing and Linting
pytest>=7.0.0
pytest-mock>=3.10.0
ruff>=0.3.3
pre-commit

# Utilities
tqdm>=4.65.0

# API and Web
fastapi>=0.110.0
uvicorn>=0.27.0
gunicorn>=21.2.0

# Dolt
doltpy

# Memory & Indexing (May move to separate requirements)
llama-index
llama-index-vector-stores-chroma
llama-index-embeddings-openai
llama-index-readers-file

# Utility & Dev
aiofiles # for async file operations in FastAPI
python-multipart # for FastAPI file uploads

# Optional/Experiment-specific (Consider separate requirements)
# tiktoken # Needed by langchain/openai usually
# pypdf # If reading PDFs directly
# unstructured # For complex document parsing
# logseq-doctor # If integrating directly with Logseq app files